requirements.txt
streamlit==1.38.0
fpdf==1.7.2
python-dotenv==1.0.1
groq==0.11.0